+++ plugins/home/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-home
 
+## 0.4.5
+
+### Patch Changes
+
+- 4a336fd292: Add name option to `createCardExtension` to remove deprecation warnings for extensions without name. Name will be required for extensions in a future release of `core-plugin-api` and therefore also in `createCardExtension`.
+- Updated dependencies
+  - @backstage/core-components@0.7.3
+  - @backstage/theme@0.2.13
+  - @backstage/core-plugin-api@0.1.13
+
 ## 0.4.4
 
 ### Patch Changes